requirements.txt
ultralytics>=8.3.233
gradio>=3.50
matplotlib
opencv-python
Pillow
tqdm
numpy
torch>=2.0 